Script now exports a csv of staff and their servers
The file that this script used to read was merged with another file, but the format of that file could still be of use. This script now exports data from the merged file into a csv organized by staff rather than server. The CSV key is staff and the value is all of the servers that staff member is responsible for.